requirements.txt
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
openai>=1.0.0
tiktoken>=0.5.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
datasets>=2.14.0
tqdm>=4.65.0
plotly>=5.14.0

